Extract inventory reporting and valuation modules
